+++ src/dreamstate/core/internals/internals.test.ts
@@ -9,7 +9,9 @@ import {
   CONTEXT_SIGNAL_HANDLERS_REGISTRY,
   CONTEXT_SIGNAL_METADATA_REGISTRY,
   CONTEXT_QUERY_METADATA_REGISTRY,
-  SIGNAL_LISTENERS_REGISTRY, CONTEXT_SERVICES_REFERENCES
+  SIGNAL_LISTENERS_REGISTRY,
+  CONTEXT_SERVICES_REFERENCES,
+  QUERY_PROVIDERS_REGISTRY
 } from "@/dreamstate/core/internals/index";
 import { unRegisterService } from "@/dreamstate/core/registry/unRegisterService";
 import { registerService } from "@/dreamstate/test-utils/registry/registerService";
@@ -20,7 +22,7 @@ describe("Dreamstate internals", () => {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     const internals = require("./internals");
 
-    expect(Object.keys(internals)).toHaveLength(12);
+    expect(Object.keys(internals)).toHaveLength(13);
   });
 
   it("Internals should be initialized", () => {
@@ -37,6 +39,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_SERVICES_REFERENCES).toBeInstanceOf(WeakMap);
     expect(CONTEXT_OBSERVERS_REGISTRY).toBeInstanceOf(WeakMap);
     expect(CONTEXT_QUERY_METADATA_REGISTRY).toBeInstanceOf(WeakMap);
+    expect(QUERY_PROVIDERS_REGISTRY).toBeInstanceOf(Map);
     expect(CONTEXT_SIGNAL_HANDLERS_REGISTRY).toBeInstanceOf(WeakMap);
     expect(CONTEXT_SIGNAL_METADATA_REGISTRY).toBeInstanceOf(WeakMap);
     expect(CONTEXT_STATES_REGISTRY).toBeInstanceOf(WeakMap);
@@ -52,6 +55,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_SIGNAL_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
     expect(CONTEXT_SIGNAL_HANDLERS_REGISTRY.get(TestContextManager)).toBeUndefined();
     expect(CONTEXT_QUERY_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
+    expect(QUERY_PROVIDERS_REGISTRY.size).toBe(0);
     expect(CONTEXT_SERVICES_REFERENCES.get(TestContextManager)).toBeUndefined();
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(0);
     expect(CONTEXT_SERVICES_ACTIVATED.size).toBe(0);
@@ -68,6 +72,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_QUERY_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined(); // No query listeners here.
     expect(CONTEXT_SERVICES_REFERENCES.get(TestContextManager)).toBe(0);
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(1);
+    expect(QUERY_PROVIDERS_REGISTRY.size).toBe(0);
     expect(CONTEXT_SERVICES_ACTIVATED.size).toBe(1);
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeUndefined();
 
@@ -81,6 +86,7 @@ describe("Dreamstate internals", () => {
     expect(CONTEXT_QUERY_METADATA_REGISTRY.get(TestContextManager)).toBeUndefined();
     expect(CONTEXT_SERVICES_REFERENCES.get(TestContextManager)).toBe(0);
     expect(SIGNAL_LISTENERS_REGISTRY.size).toBe(0);
+    expect(QUERY_PROVIDERS_REGISTRY.size).toBe(0);
     expect(CONTEXT_SERVICES_ACTIVATED.size).toBe(0);
     expect(CONTEXT_REACT_CONTEXTS_REGISTRY.get(TestContextManager)).toBeUndefined();
   });