+++ src/dreamstate/core/queries/querySingle.ts
@@ -1,10 +1,18 @@
+import { ContextService } from "@/dreamstate";
 import {
   CONTEXT_QUERY_METADATA_REGISTRY,
   CONTEXT_SERVICES_ACTIVATED,
-  CONTEXT_SERVICES_REGISTRY
+  CONTEXT_SERVICES_REGISTRY,
+  QUERY_PROVIDERS_REGISTRY
 } from "@/dreamstate/core/internals";
 import { promisifyQuery } from "@/dreamstate/core/queries/promisifyQuery";
-import { IOptionalQueryRequest, TOptionalQueryResponse, TQueryType } from "@/dreamstate/types";
+import {
+  IOptionalQueryRequest,
+  TAnyContextServiceConstructor,
+  TOptionalQueryResponse,
+  TQueryListener,
+  TQueryType
+} from "@/dreamstate/types";
 
 /**
  * Get single query resolver from possible metadata.
@@ -20,10 +28,25 @@ export function querySingle<
     if (CONTEXT_QUERY_METADATA_REGISTRY.has(service) && CONTEXT_SERVICES_REGISTRY.has(service)) {
       for (const [ method, type ] of CONTEXT_QUERY_METADATA_REGISTRY.get(service)!) {
         if (type === query.type) {
-          return promisifyQuery(CONTEXT_SERVICES_REGISTRY.get(service)!, method, query);
+          const handlerService: ContextService = CONTEXT_SERVICES_REGISTRY.get(service)!;
+
+          return promisifyQuery(
+            (handlerService as any)[method].bind(handlerService),
+            query,
+            handlerService.constructor as TAnyContextServiceConstructor
+          );
         }
       }
     }
+  }
+
+  /**
+   * From class providers fallback to manually listed query provider factories.
+   */
+  if (QUERY_PROVIDERS_REGISTRY.has(query.type)) {
+    const handlerFunction: TQueryListener<T, D> = QUERY_PROVIDERS_REGISTRY.get(query.type)![0];
+
+    return promisifyQuery(handlerFunction, query, null);
   }
 
   return Promise.resolve(null);