+++ src/dreamstate/core/signals/unsubscribeFromSignals.ts
@@ -1,12 +1,12 @@
 import { SIGNAL_LISTENERS_REGISTRY } from "@/dreamstate/core/internals";
-import { TSignalListener } from "@/dreamstate/types";
+import { TSignalListener, TSignalType } from "@/dreamstate/types";
 
 /**
- * Unsubscribe from all signals and listen everything.
+ * Unsubscribe from all signals for specified listener.
  * Not intended to be used as core feature, just for some elegant decisions.
  */
 export function unsubscribeFromSignals(
-  listener: TSignalListener<any, any>
+  listener: TSignalListener<TSignalType, any>
 ): void {
   SIGNAL_LISTENERS_REGISTRY.delete(listener);
 }