+++ src/dreamstate/index.ts
@@ -36,6 +36,8 @@ export { emitSignal } from "@/dreamstate/core/signals/emitSignal";
 export { OnQuery } from "@/dreamstate/core/queries/OnQuery";
 export { queryData } from "@/dreamstate/core/queries/queryData";
 export { queryDataSync } from "@/dreamstate/core/queries/queryDataSync";
+export { registerQueryProvider } from "@/dreamstate/core/queries/registerQueryProvider";
+export { unRegisterQueryProvider } from "@/dreamstate/core/queries/unRegisterQueryProvider";
 
 export { useManager } from "@/dreamstate/core/consumption/useManager";
 export { Consume } from "@/dreamstate/core/consumption/Consume";