+++ src/dreamstate/types/general.ts
@@ -8,4 +8,6 @@ export type TConstructor<T> = {
 
 export type TCallable = () => void;
 
+export type TAnyCallable = (...args: Array<any>) => any;
+
 export type TAnyObject = Record<string, any>;